Bug# 1614 Switching to double as the default floating point storage to match Java.  Only storing values as long if they must be stored that way.

+++ Etomo/src/etomo/storage/LogFile.java
@@ -56,8 +56,8 @@ import etomo.util.Utilities;
 public final class LogFile {
   public static final String rcsid = "$Id$";
 
-  private static final long NO_ID = -1;
-  public static final long NO_WAIT_LIMIT = -1;
+  private static final int NO_ID = -1;
+  public static final int NO_WAIT_LIMIT = -1;
   private static final String PUBLIC_EXCEPTION_MESSAGE = "\nPlease make a copy "
       + "of the current etomo_err.log file and inform the software developer.";
 
@@ -525,7 +525,7 @@ public final class LogFile {
   }
 
   /**
-   * Run open(long) with no wait limit.  This function can cause deadlock.
+   * Run open with no wait limit.  This function can cause deadlock.
    * @see waitForLock()
    * @return
    */
@@ -1079,9 +1079,9 @@ public final class LogFile {
 
     private boolean warningDisplayed = false;
     private boolean locked = false;
-    private long currentId = NO_ID;
-    private long writeId = NO_ID;
-    private long fileId = NO_ID;
+    private int currentId = NO_ID;
+    private int writeId = NO_ID;
+    private int fileId = NO_ID;
 
     private Lock(final LogFile logFile) {
       this.logFile = logFile;
@@ -1103,7 +1103,7 @@ public final class LogFile {
       return id.toString();
     }
 
-    private static String makeKey(final long id) {
+    private static String makeKey(final int id) {
       return String.valueOf(id);
     }
 
@@ -1321,7 +1321,7 @@ public final class LogFile {
       }
     }
 
-    private long getWriteId() {
+    private int getWriteId() {
       return writeId;
     }
   }
@@ -1586,17 +1586,17 @@ public final class LogFile {
   }
 
   static class Id {
-    private long id = NO_ID;
+    private int id = NO_ID;
 
-    void set(long input) {
+    void set(int input) {
       this.id = input;
     }
 
-    long get() {
+    int get() {
       return id;
     }
 
-    private boolean equals(long input) {
+    private boolean equals(int input) {
       return id == input;
     }
 
@@ -1605,7 +1605,7 @@ public final class LogFile {
     }
 
     public String toString() {
-      return Long.toString(get());
+      return Integer.toString(get());
     }
   }
